Upgrade @twreporter/react-components to 4.0.0
